Updates to move towards single_bet module having all key single bet strategies (including Place, Pass, DontPass)
* Add two new SimpleStrategyModes for SingleBets: ADD_IF_POINT_ON, ADD_IF_POINT_OFF
* Remove BetPassAmount and BetDontPass from examples, using the renamed version in single_bet with different default mode. This solves the duplication problem and generalizes better
* Internalize _BaseSingleBet class
* Move BetPlace from examples to single_bet in strategy module
* Clean up strategy __init__ file and update any testing calls
