+++ crapssim/strategy/__init__.py
@@ -4,9 +4,19 @@ player should make, remove, or change. Strategies are applied for the player to 
 after the previous bets and table have been updated.
 """
 
-from crapssim.strategy.core import Strategy, AggregateStrategy, BetIfTrue, RemoveIfTrue, \
-    IfBetNotExist, BetPointOff, BetPointOn, CountStrategy
-from crapssim.strategy.odds import OddsMultiplierStrategy, PassLineOddsMultiplier, \
-    DontPassOddsMultiplier
-from crapssim.strategy.examples import BetPassLine, BetDontPass, \
-    BetPlace
+from crapssim.strategy.core import (
+    AggregateStrategy,
+    BetIfTrue,
+    BetPointOff,
+    BetPointOn,
+    CountStrategy,
+    IfBetNotExist,
+    RemoveIfTrue,
+    Strategy,
+)
+from crapssim.strategy.odds import (
+    DontPassOddsMultiplier,
+    OddsMultiplierStrategy,
+    PassLineOddsMultiplier,
+)
+from crapssim.strategy.single_bet import BetDontPass, BetPassLine, BetPlace
