+++ crapssim/strategy/single_bet.py
@@ -8,6 +8,7 @@ from crapssim.bet import (
     Any7,
     Bet,
     Boxcars,
+    Come,
     DontCome,
     DontPass,
     Field,
@@ -20,17 +21,121 @@ from crapssim.bet import (
     Yo,
 )
 from crapssim.strategy import BetIfTrue
-from crapssim.strategy.core import Player, Strategy
+from crapssim.strategy.core import (
+    BetPointOff,
+    BetPointOn,
+    IfBetNotExist,
+    Player,
+    RemoveIfTrue,
+    Strategy,
+)
 from crapssim.strategy.odds import OddsAmountStrategy
 
 
 class SimpleStrategyMode(enum.Enum):
     ADD_IF_NON_EXISTENT = enum.auto()
+    ADD_IF_POINT_OFF = enum.auto()
+    ADD_IF_POINT_ON = enum.auto()
     ADD_OR_INCREASE = enum.auto()
     REPLACE = enum.auto()
 
 
-class BaseSingleBet(Strategy):
+class BetPlace(Strategy):
+    """Strategy that makes multiple Place bets of given amounts. It can also skip making the bet
+    if the point is the same as the given bet number."""
+
+    # TODO: add bet mode to this (default currently is point off)
+    def __init__(
+        self,
+        place_bet_amounts: dict[int, float],
+        skip_point: bool = True,
+        skip_come: bool = False,
+    ):
+        """Strategy for making multiple place bets.
+
+        Parameters
+        ----------
+        place_bet_amounts
+            Dictionary of the point to make the Place bet on and the amount of the
+            place bet to make.
+        skip_point
+            If True don't make the bet on the given Place if that's the number the tables Point
+            is on.
+        skip_come
+            If True don't make the bet on the given Place if there is a Come bet with that Point
+            already on that number.
+        """
+        super().__init__()
+        self.place_bet_amounts = place_bet_amounts
+        self.skip_point = skip_point
+        self.skip_come = skip_come
+
+    def completed(self, player: Player) -> bool:
+        """The strategy is completed if the player can no longer make any of the place bets in the
+        place_bet_amounts dictionary and there are no Place bets on the table.
+
+        Parameters
+        ----------
+        player
+            The player to check the bankroll for
+
+        Returns
+        -------
+        True if there are no Place bets on the table and the player can't make any more Place bets
+        because their bankroll is too low.
+        """
+        return (
+            player.bankroll < min(x for x in self.place_bet_amounts.values())
+            and len([x for x in player.bets if isinstance(x, Place)]) == 0
+        )
+
+    def update_bets(self, player: Player) -> None:
+        """Add the place bets on the numbers and amounts defined by place_bet_amounts.
+
+        Parameters
+        ----------
+        player
+            The player to add the place bets to.
+        """
+        if self.skip_point:
+            self.remove_point_bet(player)
+
+        for number, amount in self.place_bet_amounts.items():
+            if self.skip_point and number == player.table.point.number:
+                continue
+            if player.table.point.status == "Off":
+                continue
+            if self.skip_come:
+                come_numbers = [
+                    x.point.number for x in player.bets if isinstance(x, Come)
+                ]
+                if number in come_numbers:
+                    continue
+            IfBetNotExist(Place(number, amount)).update_bets(player)
+
+    @staticmethod
+    def remove_point_bet(player: Player) -> None:
+        """If skip_point is true and the player has a place bet for the table point number,
+        remove the Place bet.
+
+        Parameters
+        ----------
+        player
+            The player to check and see if they have the given bet.
+        """
+        point = player.table.point.number
+        RemoveIfTrue(
+            lambda b, p: isinstance(b, Place) and b.number == point
+        ).update_bets(player)
+
+    def __repr__(self) -> str:
+        return (
+            f"{self.__class__.__name__}(place_bet_amounts={self.place_bet_amounts},"
+            f" skip_point={self.skip_point}, skip_come={self.skip_come})"
+        )
+
+
+class _BaseSingleBet(Strategy):
     def __init__(
         self,
         bet: Bet,
@@ -49,6 +154,10 @@ class BaseSingleBet(Strategy):
 
         if self.mode == SimpleStrategyMode.ADD_IF_NON_EXISTENT:
             BetIfTrue(self.bet, lambda p: self.bet not in p.bets).update_bets(player)
+        elif self.mode == SimpleStrategyMode.ADD_IF_POINT_ON:
+            BetPointOn(self.bet).update_bets(player)
+        elif self.mode == SimpleStrategyMode.ADD_IF_POINT_OFF:
+            BetPointOff(self.bet).update_bets(player)
         elif self.mode == SimpleStrategyMode.ADD_OR_INCREASE:
             player.add_bet(self.bet)
         elif self.mode == SimpleStrategyMode.REPLACE:
@@ -58,13 +167,22 @@ class BaseSingleBet(Strategy):
             player.add_bet(self.bet)
 
 
-class PassLineAmount(BaseSingleBet):
+class BetPassLine(_BaseSingleBet):
     def __init__(
         self,
         bet_amount: typing.SupportsFloat,
-        mode=SimpleStrategyMode.ADD_IF_NON_EXISTENT,
+        mode=SimpleStrategyMode.ADD_IF_POINT_OFF,
     ):
         super().__init__(PassLine(bet_amount), mode=mode)
+
+
+class BetDontPass(_BaseSingleBet):
+    def __init__(
+        self,
+        bet_amount: typing.SupportsFloat,
+        mode=SimpleStrategyMode.ADD_IF_POINT_OFF,
+    ):
+        super().__init__(DontPass(bet_amount), mode=mode)
 
 
 class PassLineOddsAmount(OddsAmountStrategy):
@@ -76,15 +194,6 @@ class PassLineOddsAmount(OddsAmountStrategy):
         super().__init__(PassLine, {x: bet_amount for x in numbers})
 
 
-class DontPassAmount(BaseSingleBet):
-    def __init__(
-        self,
-        bet_amount: typing.SupportsFloat,
-        mode=SimpleStrategyMode.ADD_IF_NON_EXISTENT,
-    ):
-        super().__init__(DontPass(bet_amount), mode=mode)
-
-
 class DontPassOddsAmount(OddsAmountStrategy):
     def __init__(
         self,
@@ -94,7 +203,7 @@ class DontPassOddsAmount(OddsAmountStrategy):
         super().__init__(DontPass, {x: bet_amount for x in numbers})
 
 
-class ComeAmount(BaseSingleBet):
+class ComeAmount(_BaseSingleBet):
     def __init__(
         self,
         bet_amount: typing.SupportsFloat,
@@ -112,7 +221,7 @@ class ComeOddsAmount(OddsAmountStrategy):
         super().__init__(DontPass, {x: bet_amount for x in numbers})
 
 
-class DontComeAmount(BaseSingleBet):
+class DontComeAmount(_BaseSingleBet):
     def __init__(
         self,
         bet_amount: typing.SupportsFloat,
@@ -130,7 +239,7 @@ class DontComeOddsAmount(OddsAmountStrategy):
         super().__init__(DontCome, {x: bet_amount for x in numbers})
 
 
-class BetHardWay(BaseSingleBet):
+class BetHardWay(_BaseSingleBet):
     def __init__(
         self,
         number: tuple[int],
@@ -142,7 +251,7 @@ class BetHardWay(BaseSingleBet):
         super().__init__(HardWay(number, bet_amount), mode=mode)
 
 
-class FieldAmount(BaseSingleBet):
+class FieldAmount(_BaseSingleBet):
     def __init__(
         self,
         bet_amount: typing.SupportsFloat,
@@ -151,7 +260,7 @@ class FieldAmount(BaseSingleBet):
         super().__init__(Field(bet_amount), mode=mode)
 
 
-class BetAny7(BaseSingleBet):
+class BetAny7(_BaseSingleBet):
     def __init__(
         self,
         bet_amount: typing.SupportsFloat,
@@ -160,7 +269,7 @@ class BetAny7(BaseSingleBet):
         super().__init__(Any7(bet_amount), mode=mode)
 
 
-class BetTwo(BaseSingleBet):
+class BetTwo(_BaseSingleBet):
     def __init__(
         self,
         bet_amount: typing.SupportsFloat,
@@ -169,7 +278,7 @@ class BetTwo(BaseSingleBet):
         super().__init__(Two(bet_amount), mode=mode)
 
 
-class BetThree(BaseSingleBet):
+class BetThree(_BaseSingleBet):
     def __init__(
         self,
         bet_amount: typing.SupportsFloat,
@@ -178,7 +287,7 @@ class BetThree(BaseSingleBet):
         super().__init__(Three(bet_amount), mode=mode)
 
 
-class BetYo(BaseSingleBet):
+class BetYo(_BaseSingleBet):
     def __init__(
         self,
         bet_amount: typing.SupportsFloat,
@@ -187,7 +296,7 @@ class BetYo(BaseSingleBet):
         super().__init__(Yo(bet_amount), mode=mode)
 
 
-class BetBoxcars(BaseSingleBet):
+class BetBoxcars(_BaseSingleBet):
     def __init__(
         self,
         bet_amount: typing.SupportsFloat,
@@ -196,6 +305,6 @@ class BetBoxcars(BaseSingleBet):
         super().__init__(Boxcars(bet_amount), mode=mode)
 
 
-class BetFire(BaseSingleBet):
+class BetFire(_BaseSingleBet):
     def __init__(self, bet_amount: float, mode=SimpleStrategyMode.ADD_IF_NON_EXISTENT):
         super().__init__(Fire(bet_amount), mode=mode)
