+++ tests/unit/test_strategy.py
@@ -40,7 +40,7 @@ from crapssim.strategy.examples import (
     TwoCome,
 )
 from crapssim.strategy.odds import OddsAmountStrategy, OddsMultiplierStrategy
-from crapssim.strategy.single_bet import BaseSingleBet, SimpleStrategyMode
+from crapssim.strategy.single_bet import SimpleStrategyMode, _BaseSingleBet
 
 
 @pytest.fixture
@@ -538,14 +538,14 @@ def test_odds_multiplier_dont_come_bet_placed(player):
 
 
 def test_base_single_bet_add_if_non_existent_add(player):
-    strategy = BaseSingleBet(PassLine(5))
+    strategy = _BaseSingleBet(PassLine(5))
     player.add_bet = MagicMock()
     strategy.update_bets(player)
     player.add_bet.assert_called_with(PassLine(5))
 
 
 def test_base_single_bet_add_if_non_existent_dont_add(player):
-    strategy = BaseSingleBet(PassLine(5))
+    strategy = _BaseSingleBet(PassLine(5))
     player.add_bet = MagicMock()
     player.bets = [PassLine(5)]
     strategy.update_bets(player)
@@ -555,14 +555,14 @@ def test_base_single_bet_add_if_non_existent_dont_add(player):
 def test_base_single_bet_is_not_allowed(player):
     bet = MagicMock(PassLine)
     bet.is_allowed = MagicMock(return_value=False)
-    strategy = BaseSingleBet(bet)
+    strategy = _BaseSingleBet(bet)
     strategy.update_bets(player)
     player.add_bet = MagicMock()
     player.add_bet.assert_not_called()
 
 
 def test_base_single_bet_add_or_increase(player):
-    strategy = BaseSingleBet(PassLine(5), SimpleStrategyMode.ADD_OR_INCREASE)
+    strategy = _BaseSingleBet(PassLine(5), SimpleStrategyMode.ADD_OR_INCREASE)
     player.bets = [PassLine(5)]
     player.add_bet = MagicMock()
     strategy.update_bets(player)
@@ -570,7 +570,7 @@ def test_base_single_bet_add_or_increase(player):
 
 
 def test_base_single_bet_replace(player):
-    strategy = BaseSingleBet(PassLine(5), SimpleStrategyMode.REPLACE)
+    strategy = _BaseSingleBet(PassLine(5), SimpleStrategyMode.REPLACE)
     player.bets = [PassLine(5)]
     player.add_bet = MagicMock()
     player.remove_bet = MagicMock()
